# internal/api/client/api/swagger.yaml
openapi: 3.0.3
info:
  title: Frank Auth Server API
  description: "Frank is a comprehensive authentication and authorization server that\
    \ provides OAuth2, passwordless login, MFA, \npasskeys, SSO, enterprise SSO, webhooks,\
    \ organizations, and API keys for machine-to-machine authentication.\n"
  contact:
    name: Frank Auth Server
    url: https://github.com/juicycleff/frank
  version: 1.0.0
servers:
- url: https://auth.example.com/api/v1
  description: Production server
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
...
  /organizations:
    get:
      tags:
      - Organizations
      summary: List organizations
      description: List organizations with pagination and filtering
      operationId: listOrganizations
      parameters:
      - name: offset
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 20
      - name: search
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_23"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    post:
      tags:
      - Organizations
      summary: Create organization
      description: Create a new organization
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/organizations_body"
        required: true
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Organization already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /organizations/{id}:
    get:
      tags:
      - Organizations
      summary: Get organization
      description: Get a specific organization
      operationId: getOrganization
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    put:
      tags:
      - Organizations
      summary: Update organization
      description: Update a specific organization
      operationId: updateOrganization
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/organizations_id_body"
        required: true
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Organizations
      summary: Delete organization
      description: Delete a specific organization
      operationId: deleteOrganization
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Organization deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /organizations/{id}/members:
    get:
      tags:
      - Organizations
      summary: List organization members
      description: List members of an organization
      operationId: listOrganizationMembers
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: offset
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 20
      - name: search
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: List of organization members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_22"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    post:
      tags:
      - Organizations
      summary: Add organization member
      description: Add a user to an organization
      operationId: addOrganizationMember
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/id_members_body"
        required: true
      responses:
        "200":
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_24"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User is already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /organizations/{id}/members/{userId}:
    put:
      tags:
      - Organizations
      summary: Update organization member
      description: Update a member's roles in an organization
      operationId: updateOrganizationMember
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/members_userId_body"
        required: true
      responses:
        "200":
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_25"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Organizations
      summary: Remove organization member
      description: Remove a user from an organization
      operationId: removeOrganizationMember
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: userId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Member removed successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /organizations/{id}/features:
    get:
      tags:
      - Organizations
      summary: List organization features
      description: List enabled features for an organization
      operationId: listOrganizationFeatures
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: List of organization features
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_26"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    post:
      tags:
      - Organizations
      summary: Enable organization feature
      description: Enable a feature for an organization
      operationId: enableOrganizationFeature
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/id_features_body"
        required: true
      responses:
        "200":
          description: Feature enabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_27"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or feature not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
  /organizations/{id}/features/{featureKey}:
    delete:
      tags:
      - Organizations
      summary: Disable organization feature
      description: Disable a feature for an organization
      operationId: disableOrganizationFeature
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: featureKey
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Feature disabled successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Organization or feature not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error_error"
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        logo_url:
          type: string
          format: uri
        plan:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        owner_id:
          type: string
      example:
        metadata: ""
        updated_at: 2000-01-23T04:56:07.000+00:00
        logo_url: http://example.com/aeiou
        owner_id: owner_id
        domain: domain
        name: name
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        plan: plan
        slug: slug
...
    inline_response_200_23:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
        pagination:
          $ref: "#/components/schemas/inline_response_200_22_pagination"
      example:
        total: 0
        pagination:
          total: 5
          offset: 6
          limit: 1
        data:
        - metadata: ""
          updated_at: 2000-01-23T04:56:07.000+00:00
          logo_url: http://example.com/aeiou
          owner_id: owner_id
          domain: domain
          name: name
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          plan: plan
          slug: slug
        - metadata: ""
          updated_at: 2000-01-23T04:56:07.000+00:00
          logo_url: http://example.com/aeiou
          owner_id: owner_id
          domain: domain
          name: name
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          plan: plan
          slug: slug
    organizations_body:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        logo_url:
          type: string
          format: uri
        plan:
          type: string
        metadata:
          type: object
          additionalProperties: true
        trial_days:
          type: integer
          default: 0
        features:
          type: array
          items:
            type: string
    organizations_id_body:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        logo_url:
          type: string
          format: uri
        plan:
          type: string
        active:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
    id_members_body:
      required:
      - roles
      - user_id
      type: object
      properties:
        user_id:
          type: string
        roles:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from the /auth/login endpoint
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      description: API key for machine-to-machine authentication
      name: X-API-Key
      in: header